Refactoring 🔨
